% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar-precision.r
\name{calendar-precision}
\alias{calendar-precision}
\alias{calendar_precision.datetimeoffset}
\alias{calendar_narrow}
\alias{calendar_narrow.datetimeoffset}
\alias{calendar_narrow.default}
\alias{calendar_widen}
\alias{calendar_widen.datetimeoffset}
\alias{calendar_widen.default}
\title{Calendar precision}
\usage{
\method{calendar_precision}{datetimeoffset}(x)

calendar_narrow(x, precision)

\method{calendar_narrow}{datetimeoffset}(x, precision)

\method{calendar_narrow}{default}(x, precision)

calendar_widen(x, precision, ...)

\method{calendar_widen}{datetimeoffset}(
  x,
  precision,
  ...,
  year = 0L,
  month = 1L,
  day = 1L,
  hour = 0L,
  minute = 0L,
  second = 0L,
  nanosecond = 0L
)

\method{calendar_widen}{default}(x, precision, ...)
}
\arguments{
\item{x}{A \code{\link[=datetimeoffset]{datetimeoffset()}} vector}

\item{precision}{Precision to narrow/widen to.  Either "year", "month", "day", "hour", "minute", "second", or "nanosecond".}

\item{...}{Used by certain methods}

\item{year}{If missing what year to assume}

\item{month}{If missing what month to assume}

\item{day}{If missing what day to assume}

\item{hour}{If missing what hour to assume}

\item{minute}{If missing what minute to assume}

\item{second}{If missing what second to assume}

\item{nanosecond}{If missing what nanosecond to assume}
}
\value{
\code{calendar_precision()} returns a character vector of precisions ("year", "month", "day", "hour", "minute", "second", or "nanosecond").
\code{calendar_narrow()} and \code{calendar_widen()} return a \code{\link[=datetimeoffset]{datetimeoffset()}} vector.
}
\description{
\code{calendar_precision()} returns the "precision" of a \code{\link[=datetimeoffset]{datetimeoffset()}} vector's datetimes.
\code{calendar_widen()} sets a floor on the minimum "precision" in the vector by setting any missing
elements to their minimum possible value.
\code{calendar_narrow()} sets a cap on the maximum "precision" by setting
any more precise elements missing.
}
\examples{
  dts <- as_datetimeoffset(c("2020", "2020-04-10", "2020-04-10T10:10"))
  library("clock", exclude = c("calendar_narrow", "calendar_widen"))
  calendar_precision(dts)
  calendar_narrow(dts, "day")
  calendar_widen(dts, "day")
}
\seealso{
\code{\link[clock:calendar_precision]{clock::calendar_precision()}}, \code{\link[clock:calendar_narrow]{clock::calendar_narrow()}}, and \code{\link[clock:calendar_widen]{clock::calendar_widen()}}
}
