% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar-precision.r
\name{calendar_precision}
\alias{calendar_precision}
\alias{calendar_precision.datetimeoffset}
\title{Calendar precision}
\usage{
calendar_precision(x, ...)

\method{calendar_precision}{datetimeoffset}(x, range = FALSE, ...)
}
\arguments{
\item{x}{A \code{\link[=datetimeoffset]{datetimeoffset()}} vector}

\item{...}{Reserved for other methods.}

\item{range}{If \code{TRUE} return just the minimum and maximum "precision".}
}
\value{
\code{calendar_precision()} returns a character vector of precisions ("year", "month", "day", "hour", "minute", "second", or "nanosecond").
}
\description{
\code{calendar_precision()} returns the "precision" of a \code{\link[=datetimeoffset]{datetimeoffset()}} vector's datetimes.
}
\examples{
  dts <- as_datetimeoffset(c("2020", "2020-04-10", "2020-04-10T10:10"))
  library("clock", exclude = c("calendar_narrow", "calendar_precision", "calendar_widen"))
  calendar_precision(dts)
  calendar_precision(dts, range = TRUE)
}
\seealso{
\code{\link[clock:calendar_precision]{clock::calendar_precision()}}
}
