% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset-clock.r
\name{clock-setters}
\alias{clock-setters}
\alias{set_year.datetimeoffset}
\alias{set_month.datetimeoffset}
\alias{set_day.datetimeoffset}
\alias{set_hour.datetimeoffset}
\alias{set_minute.datetimeoffset}
\alias{set_second.datetimeoffset}
\alias{set_nanosecond.datetimeoffset}
\alias{set_zone}
\alias{set_zone.datetimeoffset}
\alias{set_zone.default}
\alias{set_hour_offset}
\alias{set_hour_offset.datetimeoffset}
\alias{set_hour_offset.default}
\alias{set_minute_offset}
\alias{set_minute_offset.datetimeoffset}
\alias{set_minute_offset.default}
\title{Set datetime components with "clock" style setters}
\usage{
\method{set_year}{datetimeoffset}(x, value, ...)

\method{set_month}{datetimeoffset}(x, value, ...)

\method{set_day}{datetimeoffset}(x, value, ...)

\method{set_hour}{datetimeoffset}(x, value, ...)

\method{set_minute}{datetimeoffset}(x, value, ...)

\method{set_second}{datetimeoffset}(x, value, ...)

\method{set_nanosecond}{datetimeoffset}(x, value, ...)

set_zone(x, value, ...)

\method{set_zone}{datetimeoffset}(x, value, ...)

\method{set_zone}{default}(x, value, ...)

set_hour_offset(x, value, ...)

\method{set_hour_offset}{datetimeoffset}(x, value, ...)

\method{set_hour_offset}{default}(x, value, ...)

set_minute_offset(x, value, ...)

\method{set_minute_offset}{datetimeoffset}(x, value, ...)

\method{set_minute_offset}{default}(x, value, ...)
}
\arguments{
\item{x}{A \code{\link[=datetimeoffset]{datetimeoffset()}} object.}

\item{value}{The replacement value.}

\item{...}{Currently ignored.}
}
\value{
A \code{\link[=datetimeoffset]{datetimeoffset()}} object.
}
\description{
\code{clock} style setter methods for \code{\link[=datetimeoffset]{datetimeoffset()}} objects
}
\details{
We implement the following datetime component setter functions from \code{clock}:
\itemize{
\item \code{\link[clock:clock-setters]{clock::set_year()}}
\item \code{\link[clock:clock-setters]{clock::set_month()}}
\item \code{\link[clock:clock-setters]{clock::set_day()}}
\item \code{\link[clock:clock-setters]{clock::set_hour()}}
\item \code{\link[clock:clock-setters]{clock::set_minute()}}
\item \code{\link[clock:clock-setters]{clock::set_second()}}
\item \code{\link[clock:clock-setters]{clock::set_nanosecond()}}
}

We also implement new \code{clock} "style" getter functions (not in \code{clock}):
\itemize{
\item \code{set_hour_offset()}
\item \code{set_minute_offset()}
\item \code{set_zone()} (changes system time but not clock time)
}
}
\examples{
library("clock", quietly = TRUE)
dt <- NA_datetimeoffset_
dt <- set_year(dt, 1918L)
dt <- set_month(dt, 11L)
dt <- set_day(dt, 11L)
dt <- set_hour(dt, 11L)
dt <- set_minute(dt, 11L)
dt <- set_second(dt, 11L)
dt <- set_nanosecond(dt, NA_integer_)
dt <- set_zone(dt, "Europe/Paris")
dt <- set_hour_offset(dt, 0L)
dt <- set_minute_offset(dt, 0L)
format(dt)
}
\seealso{
\link[clock]{clock-setters}
}
