% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision.r
\name{datetime_cast}
\alias{datetime_cast}
\alias{datetime_narrow}
\alias{datetime_narrow.datetimeoffset}
\alias{datetime_narrow.clock_calendar}
\alias{datetime_narrow.clock_time_point}
\alias{datetime_widen}
\alias{datetime_widen.datetimeoffset}
\alias{datetime_widen.clock_calendar}
\alias{datetime_widen.clock_time_point}
\title{Widen/narrow datetime precision}
\usage{
datetime_narrow(x, precision, ...)

\method{datetime_narrow}{datetimeoffset}(x, precision, ...)

\method{datetime_narrow}{clock_calendar}(x, precision, ...)

\method{datetime_narrow}{clock_time_point}(x, precision, ...)

datetime_widen(x, precision, ...)

\method{datetime_widen}{datetimeoffset}(
  x,
  precision,
  ...,
  year = 0L,
  month = 1L,
  day = 1L,
  hour = 0L,
  minute = 0L,
  second = 0L,
  nanosecond = 0L
)

\method{datetime_widen}{clock_calendar}(x, precision, ...)

\method{datetime_widen}{clock_time_point}(x, precision, ...)
}
\arguments{
\item{x}{A datetime vector.  Either \code{\link[=datetimeoffset]{datetimeoffset()}}, a "clock" "calendar", or a "clock" "time point".}

\item{precision}{Precision to narrow/widen to.  Either "year", "month", "day", "hour", "minute", "second", or "nanosecond".}

\item{...}{Used by certain methods}

\item{year}{If missing what year to assume}

\item{month}{If missing what month to assume}

\item{day}{If missing what day to assume}

\item{hour}{If missing what hour to assume}

\item{minute}{If missing what minute to assume}

\item{second}{If missing what second to assume}

\item{nanosecond}{If missing what nanosecond to assume}
}
\value{
A datetime vector.
}
\description{
\code{datetime_widen()} sets a floor on the minimum "precision" in the datetime vector
by setting any missing elements to their minimum possible value.
\code{datetime_narrow()} sets a cap on the maximum "precision" by setting
any more precise elements missing.
}
\examples{
  dts <- as_datetimeoffset(c("2020", "2020-04-10", "2020-04-10T10:10"))
  datetime_precision(dts)
  datetime_narrow(dts, "day")
  datetime_widen(dts, "day")
  datetime_widen(dts, "day", month = 6, day = 15)

  library("clock")
  ymd <- year_month_day(1918, 11, 11, 11)
  datetime_narrow(ymd, "day")
  datetime_widen(ymd, "second")

  nt <- as_naive_time(ymd)
  datetime_narrow(nt, "day")
  datetime_widen(nt, "second")
}
