% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetimeoffset.r
\docType{data}
\name{datetimeoffset_utilities}
\alias{datetimeoffset_utilities}
\alias{is_datetimeoffset}
\alias{NA_datetimeoffset_}
\title{Various "datetimeoffset" object utilities}
\format{
An object of class \code{datetimeoffset} (inherits from \code{vctrs_rcrd}, \code{vctrs_vctr}) of length 1.
}
\usage{
is_datetimeoffset(x)

NA_datetimeoffset_
}
\arguments{
\item{x}{An object to be tested}
}
\description{
\code{is_datetimeoffset()} tests whether a datetime object is of the "datetimeoffset" class.
\code{NA_datetimeoffset_} provides a "missing" "datetimeoffset" object.
}
\examples{
  is_datetimeoffset(as_datetimeoffset(Sys.time()))
  is_datetimeoffset(Sys.time())

  is.na(NA_datetimeoffset_)
  is.na(as_datetimeoffset(""))

}
\keyword{datasets}
