% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration.r
\name{duration}
\alias{duration}
\alias{vec_arith.datetimeoffset}
\alias{vec_arith.datetimeoffset.default}
\alias{vec_arith.datetimeoffset.datetimeoffset}
\alias{vec_arith.datetimeoffset.numeric}
\alias{vec_arith.numeric.datetimeoffset}
\alias{vec_arith.datetimeoffset.difftime}
\alias{vec_arith.difftime.datetimeoffset}
\alias{+,datetimeoffset,nanoduration-method}
\alias{-,datetimeoffset,nanoduration-method}
\alias{+,nanoduration,datetimeoffset-method}
\alias{+,datetimeoffset,Duration-method}
\alias{-,datetimeoffset,Duration-method}
\alias{+,Duration,datetimeoffset-method}
\title{Add/subtract datetime durations}
\usage{
\method{vec_arith}{datetimeoffset}(op, x, y, ...)

\method{vec_arith.datetimeoffset}{default}(op, x, y, ...)

\method{vec_arith.datetimeoffset}{datetimeoffset}(op, x, y, ...)

\method{vec_arith.datetimeoffset}{numeric}(op, x, y, ...)

\method{vec_arith.numeric}{datetimeoffset}(op, x, y, ...)

\method{vec_arith.datetimeoffset}{difftime}(op, x, y, ...)

\method{vec_arith.difftime}{datetimeoffset}(op, x, y, ...)

\S4method{+}{datetimeoffset,nanoduration}(e1, e2)

\S4method{-}{datetimeoffset,nanoduration}(e1, e2)

\S4method{+}{nanoduration,datetimeoffset}(e1, e2)

\S4method{+}{datetimeoffset,Duration}(e1, e2)

\S4method{-}{datetimeoffset,Duration}(e1, e2)

\S4method{+}{Duration,datetimeoffset}(e1, e2)
}
\arguments{
\item{op, x, y}{See \code{\link[vctrs:vec_arith]{vctrs::vec_arith()}}.}

\item{...}{Ignored}

\item{e1, e2}{See \link[methods:S4groupGeneric]{methods::Ops}.}
}
\description{
We can add/substract \code{\link[nanotime:nanoduration]{nanotime::nanoduration()}} or \code{\link[lubridate:duration]{lubridate::duration()}} objects with \code{\link[=datetimeoffset]{datetimeoffset()}} objects.
The difference of two \code{\link[=datetimeoffset]{datetimeoffset()}} objects is a \code{\link[nanotime:nanoduration]{nanotime::nanoduration()}} object.
One can use \code{\link[vctrs:vec_arith]{vctrs::vec_arith()}} to add/subtract \code{\link[base:difftime]{base::difftime()}} objects with \code{\link[=datetimeoffset]{datetimeoffset()}} objects.
}
\examples{
dt <- as_datetimeoffset(Sys.time())

# nanotime::nanoduration() durations
class(dt - dt)
nd <- nanotime::nanoduration(hours = 3, minutes = 0, seconds = 0, nanoseconds = 0)
dt + nd
nd + dt
dt - nd

# lubridate::duration() durations
ld <- lubridate::duration(hour = 3)
dt + ld
ld + dt
dt - ld

# base::difftime() durations
bd <- as.difftime(3, units = "hours")
vctrs::vec_arith("+", dt, bd)
vctrs::vec_arith("-", dt, bd)
vctrs::vec_arith("+", bd, dt)
dt + nanotime::as.nanoduration(bd)
dt + lubridate::as.duration(bd)

# compare DST handling to periods
boundary <- as_datetimeoffset("2009-03-08 01:59:59", tz = "America/Chicago")
boundary + lubridate::days(1) # period
boundary + lubridate::ddays(1) # duration
boundary + nanotime::nanoperiod(day = 1) # period
boundary + nanotime::nanoduration(hour = 24, minute = 0, second = 0, nanosecond = 0) # duration
}
\seealso{
\link{period} for adding/subtracting period objects.
}
