% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.r
\name{format}
\alias{format}
\alias{format.datetime_offset}
\alias{format_pdfmark}
\alias{format_pdfmark.default}
\alias{format_pdfmark.datetime_offset}
\title{Format for datetime objects with possible UTC offsets}
\usage{
\method{format}{datetime_offset}(x, ...)

format_pdfmark(x, ...)

\method{format_pdfmark}{default}(x, ...)

\method{format_pdfmark}{datetime_offset}(x, ...)
}
\arguments{
\item{x}{A \code{\link[=datetime_offset]{datetime_offset()}} object.}

\item{...}{Ignored}
}
\value{
A character vector
}
\description{
\code{format()} returns an ISO 8601 datetime string with as much \strong{known} information possible.
\code{format_pdfmark()} returns a pdfmark datetime string with as much \strong{known} information possible.
}
\examples{
  # ISO 8601 datetimes
  format(as_datetime_offset("2020-05"))
  format(as_datetime_offset("2020-05-10 20:15"))
  format(as_datetime_offset("2020-05-10 20:15:05-07"))

  # pdfmark datetimes
  format_pdfmark(as_datetime_offset("2020-05"))
  format_pdfmark(as_datetime_offset("2020-05-10 20:15"))
  format_pdfmark(as_datetime_offset("2020-05-10 20:15:05-07"))
}
