% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from-datetimeoffset.r
\name{from_datetimeoffset}
\alias{from_datetimeoffset}
\alias{as.Date.datetimeoffset}
\alias{as_date.datetimeoffset}
\alias{as.nanotime,datetimeoffset-method}
\alias{as.POSIXct.datetimeoffset}
\alias{as_date_time.datetimeoffset}
\alias{as.POSIXlt.datetimeoffset}
\title{Convert to other datetime objects}
\usage{
\method{as.Date}{datetimeoffset}(x, ..., year = 1L, month = 1L, day = 1L)

\method{as_date}{datetimeoffset}(x)

\S4method{as.nanotime}{datetimeoffset}(
  from,
  ...,
  year = 1L,
  month = 1L,
  day = 1L,
  hour = 0L,
  minute = 0L,
  second = 0L,
  nanosecond = 0L,
  tz = ""
)

\method{as.POSIXct}{datetimeoffset}(
  x,
  tz = mode_tz(x),
  ...,
  year = 1L,
  month = 1L,
  day = 1L,
  hour = 0L,
  minute = 0L,
  second = 0L,
  nanosecond = 0L
)

\method{as_date_time}{datetimeoffset}(
  x,
  zone = mode_tz(x),
  ...,
  year = 1L,
  month = 1L,
  day = 1L,
  hour = 0L,
  minute = 0L,
  second = 0L,
  nanosecond = 0L
)

\method{as.POSIXlt}{datetimeoffset}(
  x,
  tz = mode_tz(x),
  ...,
  year = 1L,
  month = 1L,
  day = 1L,
  hour = 0L,
  minute = 0L,
  second = 0L,
  nanosecond = 0L
)
}
\arguments{
\item{x}{A \code{\link[=datetimeoffset]{datetimeoffset()}} object}

\item{...}{Ignored}

\item{year}{If missing what year to assume}

\item{month}{If missing what month to assume}

\item{day}{If missing what day to assume}

\item{from}{A \code{\link[=datetimeoffset]{datetimeoffset()}} object}

\item{hour}{If missing what hour to assume}

\item{minute}{If missing what minute to assume}

\item{second}{If missing what second to assume}

\item{nanosecond}{If missing what nanosecond to assume}

\item{tz, zone}{If missing and hour offset also missing what time zone to assume}
}
\description{
We provide methods to convert \code{\link[=datetimeoffset]{datetimeoffset()}} objects to
other R datetime objects:
}
\details{
We provide the following methods:
\itemize{
\item \code{as.Date()} and \code{as_date()} returns the "local" date as a \code{\link[base:Dates]{base::Date()}} object
\item \code{as.POSIXct()} and \code{as_date_time()} returns the "local" datetime as a \code{\link[base:DateTimeClasses]{base::POSIXct()}} object
\item \code{as.POSIXlt()} returns the "local" datetime as a \code{\link[base:DateTimeClasses]{base::POSIXlt()}} object
\item \code{as.nanotime()} returns the "global" datetime as a \code{\link[nanotime:nanotime]{nanotime::nanotime()}} object
}
}
\examples{
  # {base}
  as.Date(as_datetimeoffset("2020-03-05"))
  as.Date(as_datetimeoffset("2020"))
  as.Date(as_datetimeoffset("2020"), month = 6, day = 15)
  as.POSIXct(as_datetimeoffset(Sys.time()))
  as.POSIXlt(as_datetimeoffset(Sys.time()))

  # {clock}
  clock::as_date(as_datetimeoffset(Sys.Date()))
  clock::as_date_time(as_datetimeoffset(Sys.time()))

  # {nanotime}
  nanotime::as.nanotime(as_datetimeoffset(Sys.time()))
}
