% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset-lubridate.r
\name{getset_lubridate}
\alias{getset_lubridate}
\alias{date.datetime_offset}
\alias{date<-,datetime_offset-method}
\alias{year.datetime_offset}
\alias{year<-,datetime_offset-method}
\alias{month.datetime_offset}
\alias{month<-,datetime_offset-method}
\alias{mday.datetime_offset}
\alias{day<-,datetime_offset-method}
\alias{hour.datetime_offset}
\alias{hour<-,datetime_offset-method}
\alias{minute.datetime_offset}
\alias{minute<-,datetime_offset-method}
\alias{second.datetime_offset}
\alias{second<-,datetime_offset-method}
\alias{tz.datetime_offset}
\alias{tz<-}
\title{Get/set datetime components}
\usage{
\method{date}{datetime_offset}(x)

\S4method{date}{datetime_offset}(x) <- value

\method{year}{datetime_offset}(x)

\S4method{year}{datetime_offset}(x) <- value

\method{month}{datetime_offset}(x)

\S4method{month}{datetime_offset}(x) <- value

\method{mday}{datetime_offset}(x)

\S4method{day}{datetime_offset}(x) <- value

\method{hour}{datetime_offset}(x)

\S4method{hour}{datetime_offset}(x) <- value

\method{minute}{datetime_offset}(x)

\S4method{minute}{datetime_offset}(x) <- value

\method{second}{datetime_offset}(x)

\S4method{second}{datetime_offset}(x) <- value

\method{tz}{datetime_offset}(x)

tz(x) <- value
}
\arguments{
\item{x}{A \code{\link[=datetime_offset]{datetime_offset()}} object.}

\item{value}{Replacement value}
}
\description{
S3 methods for \code{\link[=datetime_offset]{datetime_offset()}} objects for the following
datetime component getter/setter functions from \code{lubridate}:
}
\details{
\itemize{
\item \code{\link[lubridate:date]{lubridate::date()}}
\item \code{\link[lubridate:year]{lubridate::year()}}
\item \code{\link[lubridate:month]{lubridate::month()}}
\item \code{\link[lubridate:day]{lubridate::mday()}} and its alias \code{\link[lubridate:day]{lubridate::day()}}
\item \code{\link[lubridate:hour]{lubridate::hour()}}
\item \code{\link[lubridate:minute]{lubridate::minute()}}
\item \code{\link[lubridate:second]{lubridate::second()}}
\item \code{\link[lubridate:tz]{lubridate::tz()}}
}
}
\examples{
library("lubridate", exclude = c("date", "force_tz", "tz<-"))
dt <- datetime_offset(1984)
year(dt) <- 1918
year(dt)
month(dt) <- 11
month(dt)
day(dt) <- 11
day(dt)
hour(dt) <- 11
hour(dt)
minute(dt) <- 11
minute(dt)
second(dt) <- 11
second(dt)
if ("Europe/Paris" \%in\% OlsonNames()) {
  tz(dt) <- "Europe/Paris"
  tz(dt)
}

}
