% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset-other.r
\name{getset_other}
\alias{getset_other}
\alias{nanosecond}
\alias{nanosecond.default}
\alias{nanosecond.datetimeoffset}
\alias{nanosecond<-}
\alias{nanosecond<-.default}
\alias{nanosecond<-.datetimeoffset}
\alias{hour_offset}
\alias{hour_offset.datetimeoffset}
\alias{hour_offset.POSIXlt}
\alias{hour_offset.POSIXct}
\alias{hour_offset.default}
\alias{hour_offset<-}
\alias{hour_offset<-.datetimeoffset}
\alias{minute_offset}
\alias{minute_offset.datetimeoffset}
\alias{minute_offset.POSIXlt}
\alias{minute_offset.POSIXct}
\alias{minute_offset.default}
\alias{minute_offset<-}
\alias{minute_offset<-.datetimeoffset}
\title{Get/set datetime components not supported by lubridate}
\usage{
nanosecond(x)

\method{nanosecond}{default}(x)

\method{nanosecond}{datetimeoffset}(x)

nanosecond(x) <- value

\method{nanosecond}{default}(x) <- value

\method{nanosecond}{datetimeoffset}(x) <- value

hour_offset(x)

\method{hour_offset}{datetimeoffset}(x)

\method{hour_offset}{POSIXlt}(x)

\method{hour_offset}{POSIXct}(x)

\method{hour_offset}{default}(x)

hour_offset(x) <- value

\method{hour_offset}{datetimeoffset}(x) <- value

minute_offset(x)

\method{minute_offset}{datetimeoffset}(x)

\method{minute_offset}{POSIXlt}(x)

\method{minute_offset}{POSIXct}(x)

\method{minute_offset}{default}(x)

minute_offset(x) <- value

\method{minute_offset}{datetimeoffset}(x) <- value
}
\arguments{
\item{x}{A \code{\link[=datetimeoffset]{datetimeoffset()}} object.}

\item{value}{Replacement value}
}
\description{
Datetime component getter/setter functions not supported by \code{lubridate}:
}
\details{
\itemize{
\item \code{nanosecond()} and \verb{nanosecond<-()} can be used to get/set nanoseconds.
\item \code{hour_offset()} and \verb{hour_offset<-()} can be used to get/set UTC offset hours
\item \code{minute_offset()} and \verb{minute_offset<-()} can be used to get/set UTC offset minutes
}
}
\examples{
  nanosecond(Sys.Date())
  dt <- as_datetimeoffset("2020-04-04T07:30:20")
  nanosecond(dt) <- 3e6 # 3 milliseconds
  nanosecond(dt)
  nanosecond(dt) <- 3e3 # 3 microseconds
  nanosecond(dt)
  nanosecond(dt) <- 3 # 3 nanoseconds
  nanosecond(dt)
  dt <- as_datetimeoffset("2020-04-08T20:12:16")
  hour_offset(dt) <- -7
  hour_offset(dt)
  minute_offset(dt) <- 30
  minute_offset(dt)
  dt <- Sys.time()
  hour_offset(dt)
  minute_offset(dt)
}
\seealso{
\link{getset_lubridate}
}
