% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanosecond.r
\name{nanosecond}
\alias{nanosecond}
\alias{nanosecond.default}
\alias{nanosecond.datetime_offset}
\alias{nanosecond<-}
\alias{nanosecond<-.default}
\alias{nanosecond<-.datetime_offset}
\title{Get/set nanoseconds}
\usage{
nanosecond(x)

\method{nanosecond}{default}(x)

\method{nanosecond}{datetime_offset}(x)

nanosecond(x) <- value

\method{nanosecond}{default}(x) <- value

\method{nanosecond}{datetime_offset}(x) <- value
}
\arguments{
\item{x}{Datetime object to get nanoseconds from}

\item{value}{Integer of nanoseconds}
}
\description{
\code{nanosecond()} and \verb{nanosecond<-()} can be used to get/set nanoseconds.
}
\examples{
  nanosecond(Sys.Date())
  dt <- as_datetime_offset("2020-04-04T07:30:20")
  nanosecond(dt) <- 3e6 # 3 milliseconds
  nanosecond(dt)
  nanosecond(dt) <- 3e3 # 3 microseconds
  nanosecond(dt)
  nanosecond(dt) <- 3 # 3 nanoseconds
  nanosecond(dt)
}
