% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.r
\name{period}
\alias{period}
\alias{+,datetimeoffset,nanoperiod-method}
\alias{-,datetimeoffset,nanoperiod-method}
\alias{+,nanoperiod,datetimeoffset-method}
\alias{+,datetimeoffset,Period-method}
\alias{-,datetimeoffset,Period-method}
\alias{+,Period,datetimeoffset-method}
\title{Add/subtract datetime periods}
\usage{
\S4method{+}{datetimeoffset,nanoperiod}(e1, e2)

\S4method{-}{datetimeoffset,nanoperiod}(e1, e2)

\S4method{+}{nanoperiod,datetimeoffset}(e1, e2)

\S4method{+}{datetimeoffset,Period}(e1, e2)

\S4method{-}{datetimeoffset,Period}(e1, e2)

\S4method{+}{Period,datetimeoffset}(e1, e2)
}
\arguments{
\item{e1, e2}{See \link[methods:S4groupGeneric]{methods::Ops}.}
}
\description{
We can add/substract \code{\link[nanotime:nanoperiod]{nanotime::nanoperiod()}} or \code{\link[lubridate:period]{lubridate::period()}} objects with \code{\link[=datetimeoffset]{datetimeoffset()}} objects.
}
\examples{
dt <- as_datetimeoffset(Sys.time())
np <- nanotime::nanoperiod(months = 12)
dt + np
np + dt
dt - np
lp <- lubridate::period(year = 1)
dt + lp
lp + dt
dt - lp

# compare DST handling to durations
boundary <- as_datetimeoffset("2009-03-08 01:59:59", tz = "America/Chicago")
boundary + lubridate::days(1) # period
boundary + lubridate::ddays(1) # duration
boundary + nanotime::nanoperiod(day = 1) # period
boundary + nanotime::nanoduration(hour = 24, minute = 0, second = 0, nanosecond = 0) # duration
}
\seealso{
\link{duration} for adding/subtracting duration objects.
}
