% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset-offsets.r
\name{utc_offsets}
\alias{utc_offsets}
\alias{hour_offset}
\alias{hour_offset.datetime_offset}
\alias{hour_offset.POSIXlt}
\alias{hour_offset.POSIXct}
\alias{hour_offset.default}
\alias{hour_offset<-}
\alias{hour_offset<-.datetime_offset}
\alias{minute_offset}
\alias{minute_offset.datetime_offset}
\alias{minute_offset.POSIXlt}
\alias{minute_offset.POSIXct}
\alias{minute_offset.default}
\alias{minute_offset<-}
\alias{minute_offset<-.datetime_offset}
\title{Get/set UTC offsets}
\usage{
hour_offset(x)

\method{hour_offset}{datetime_offset}(x)

\method{hour_offset}{POSIXlt}(x)

\method{hour_offset}{POSIXct}(x)

\method{hour_offset}{default}(x)

hour_offset(x) <- value

\method{hour_offset}{datetime_offset}(x) <- value

minute_offset(x)

\method{minute_offset}{datetime_offset}(x)

\method{minute_offset}{POSIXlt}(x)

\method{minute_offset}{POSIXct}(x)

\method{minute_offset}{default}(x)

minute_offset(x) <- value

\method{minute_offset}{datetime_offset}(x) <- value
}
\arguments{
\item{x}{Object to get UTC offsets from}

\item{value}{Replacement value}
}
\description{
\code{hour_offset()} and \code{minute_offset()} can be used to get/set UTC offsets.
}
\examples{
dt <- as_datetime_offset("2020-04-08T20:12:16")
hour_offset(dt) <- -7
hour_offset(dt)
minute_offset(dt) <- 30
minute_offset(dt)
dt <- Sys.time()
hour_offset(dt)
minute_offset(dt)
}
