% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekdays.r
\name{weekdays.datetimeoffset}
\alias{weekdays.datetimeoffset}
\alias{months.datetimeoffset}
\alias{quarters.datetimeoffset}
\alias{julian.datetimeoffset}
\title{Base Date extractors}
\usage{
\method{weekdays}{datetimeoffset}(x, abbreviate = FALSE)

\method{months}{datetimeoffset}(x, abbreviate = FALSE)

\method{quarters}{datetimeoffset}(x, ...)

\method{julian}{datetimeoffset}(x, origin = as.Date("1970-01-01"), ...)
}
\arguments{
\item{x}{A \code{\link[=datetimeoffset]{datetimeoffset()}} datetime}

\item{abbreviate}{Logical vector for whether the names should be abbreviated}

\item{...}{Ignored}

\item{origin}{Length one datetime of origin}
}
\description{
\code{\link[base:weekday.POSIXt]{base::weekdays()}}, \code{\link[base:weekday.POSIXt]{base::months()}}, \code{\link[base:weekday.POSIXt]{base::quarters()}}, and \code{\link[base:weekday.POSIXt]{base::julian()}}
are base generic functions to extract the weekday, month, quarter, or Julian time.
}
\examples{
  dto <- datetimeoffset_now()
  print(dto)
  weekdays(dto)
  months(dto)
  quarters(dto)
  julian(dto)

}
