% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset-clock.r
\name{clock-getters}
\alias{clock-getters}
\alias{get_year.datetimeoffset}
\alias{get_month.datetimeoffset}
\alias{get_day.datetimeoffset}
\alias{get_hour.datetimeoffset}
\alias{get_minute.datetimeoffset}
\alias{get_second.datetimeoffset}
\alias{get_nanosecond.datetimeoffset}
\alias{get_zone}
\alias{get_zone.datetimeoffset}
\alias{get_zone.POSIXt}
\alias{get_zone.default}
\alias{get_hour_offset}
\alias{get_hour_offset.datetimeoffset}
\alias{get_hour_offset.default}
\alias{get_hour_offset.POSIXt}
\alias{get_minute_offset}
\alias{get_minute_offset.datetimeoffset}
\alias{get_minute_offset.default}
\alias{get_minute_offset.POSIXt}
\title{Get datetime components with "clock" style getters}
\usage{
\method{get_year}{datetimeoffset}(x)

\method{get_month}{datetimeoffset}(x)

\method{get_day}{datetimeoffset}(x)

\method{get_hour}{datetimeoffset}(x)

\method{get_minute}{datetimeoffset}(x)

\method{get_second}{datetimeoffset}(x)

\method{get_nanosecond}{datetimeoffset}(x)

get_zone(x)

\method{get_zone}{datetimeoffset}(x)

\method{get_zone}{POSIXt}(x)

\method{get_zone}{default}(x)

get_hour_offset(x)

\method{get_hour_offset}{datetimeoffset}(x)

\method{get_hour_offset}{default}(x)

\method{get_hour_offset}{POSIXt}(x)

get_minute_offset(x)

\method{get_minute_offset}{datetimeoffset}(x)

\method{get_minute_offset}{default}(x)

\method{get_minute_offset}{POSIXt}(x)
}
\arguments{
\item{x}{A \code{\link[=datetimeoffset]{datetimeoffset()}} object.}
}
\value{
The component
}
\description{
\code{clock} style getter methods for \code{\link[=datetimeoffset]{datetimeoffset()}} objects
}
\details{
We implement the following datetime component getter functions from \code{clock}:
\itemize{
\item \code{\link[clock:clock-getters]{clock::get_year()}}
\item \code{\link[clock:clock-getters]{clock::get_month()}}
\item \code{\link[clock:clock-getters]{clock::get_day()}}
\item \code{\link[clock:clock-getters]{clock::get_hour()}}
\item \code{\link[clock:clock-getters]{clock::get_minute()}}
\item \code{\link[clock:clock-getters]{clock::get_second()}}
\item \code{\link[clock:clock-getters]{clock::get_nanosecond()}}
}

We also implement new \code{clock} "style" getter functions (not in \code{clock}):
\itemize{
\item \code{get_hour_offset()}
\item \code{get_minute_offset()}
\item \code{get_zone()}
}
}
\examples{
library("clock", quietly = TRUE)
if ("Europe/Paris" \%in\% OlsonNames()) {
  dt <- as_datetimeoffset("1918-11-11T11:11:11+00:00[Europe/Paris]")
} else {
  dt <- as_datetimeoffset("1918-11-11T11:11:11")
}
get_year(dt)
get_month(dt)
get_day(dt)
get_hour(dt)
get_minute(dt)
get_second(dt)
get_nanosecond(dt)
get_zone(dt)
get_hour_offset(dt)
get_minute_offset(dt)
}
\seealso{
\link[clock]{clock-getters}
}
