% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_datetime_offset.r
\name{from_datetime_offset}
\alias{from_datetime_offset}
\alias{as.Date.datetime_offset}
\title{Convert to other datetime objects}
\usage{
\method{as.Date}{datetime_offset}(x, month = 1, day = 1, ...)
}
\arguments{
\item{x}{A \code{\link[=datetime_offset]{datetime_offset()}} object}

\item{month}{If missing what month to assume}

\item{day}{If missing what day to assume}

\item{...}{Ignored}
}
\description{
We provide methods to convert \code{\link[=datetime_offset]{datetime_offset()}} objects to the following
other R datetime objects:
}
\details{
\itemize{
\item \code{as.Date()} returns the "local" date as a \code{\link[base:Dates]{base::Date()}} object
}
}
\examples{
  as.Date(as_datetime_offset("2020-03-05"))
  as.Date(as_datetime_offset("2020"))
  as.Date(as_datetime_offset("2020"), 6, 15)
}
