% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from-datetimeoffset.r
\name{from_datetimeoffset}
\alias{from_datetimeoffset}
\alias{as.Date.datetimeoffset}
\alias{as.nanotime,datetimeoffset-method}
\alias{as.POSIXct.datetimeoffset}
\alias{as.POSIXlt.datetimeoffset}
\title{Convert to other datetime objects}
\usage{
\method{as.Date}{datetimeoffset}(x, year = 1970L, month = 1L, day = 1L, ...)

\S4method{as.nanotime}{datetimeoffset}(
  from,
  ...,
  year = 1970L,
  month = 1L,
  day = 1L,
  hour = 0L,
  minute = 0L,
  second = 0L,
  nanosecond = 0L,
  tz = ""
)

\method{as.POSIXct}{datetimeoffset}(
  x,
  tz = mode_tz(x),
  ...,
  year = 1970L,
  month = 1L,
  day = 1L,
  hour = 0L,
  minute = 0L,
  second = 0L,
  nanosecond = 0L
)

\method{as.POSIXlt}{datetimeoffset}(
  x,
  tz = mode_tz(x),
  ...,
  year = 1970L,
  month = 1L,
  day = 1L,
  hour = 0L,
  minute = 0L,
  second = 0L,
  nanosecond = 0L
)
}
\arguments{
\item{x}{A \code{\link[=datetimeoffset]{datetimeoffset()}} object}

\item{year}{If missing what year to assume}

\item{month}{If missing what month to assume}

\item{day}{If missing what day to assume}

\item{...}{Ignored}

\item{from}{A \code{\link[=datetimeoffset]{datetimeoffset()}} object}

\item{hour}{If missing what hour to assume}

\item{minute}{If missing what minute to assume}

\item{second}{If missing what second to assume}

\item{nanosecond}{If missing what nanosecond to assume}

\item{tz}{If missing and hour offset also missing what time zone to assume}
}
\description{
We provide methods to convert \code{\link[=datetimeoffset]{datetimeoffset()}} objects to the following
other R datetime objects:
}
\details{
\itemize{
\item \code{as.Date()} returns the "local" date as a \code{\link[base:Dates]{base::Date()}} object
}
}
\examples{
  as.Date(as_datetimeoffset("2020-03-05"))
  as.Date(as_datetimeoffset("2020"))
  as.Date(as_datetimeoffset("2020"), month = 6, day = 15)

  as.nanotime(as_datetimeoffset(Sys.time()))
  as.POSIXct(as_datetimeoffset(Sys.time()), tz = "GMT")
}
