% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset-lubridate.r
\name{getset_lubridate}
\alias{getset_lubridate}
\alias{date}
\alias{date.datetimeoffset}
\alias{date<-,datetimeoffset-method}
\alias{year.datetimeoffset}
\alias{year<-,datetimeoffset-method}
\alias{month.datetimeoffset}
\alias{month<-,datetimeoffset-method}
\alias{mday.datetimeoffset}
\alias{day<-,datetimeoffset-method}
\alias{hour.datetimeoffset}
\alias{hour<-,datetimeoffset-method}
\alias{minute.datetimeoffset}
\alias{minute<-,datetimeoffset-method}
\alias{second.datetimeoffset}
\alias{second<-,datetimeoffset-method}
\alias{tz.datetimeoffset}
\alias{tz<-}
\title{Get/set datetime components supported by lubridate}
\usage{
date(x)

\method{date}{datetimeoffset}(x)

\S4method{date}{datetimeoffset}(x) <- value

\method{year}{datetimeoffset}(x)

\S4method{year}{datetimeoffset}(x) <- value

\method{month}{datetimeoffset}(x)

\S4method{month}{datetimeoffset}(x) <- value

\method{mday}{datetimeoffset}(x)

\S4method{day}{datetimeoffset}(x) <- value

\method{hour}{datetimeoffset}(x)

\S4method{hour}{datetimeoffset}(x) <- value

\method{minute}{datetimeoffset}(x)

\S4method{minute}{datetimeoffset}(x) <- value

\method{second}{datetimeoffset}(x)

\S4method{second}{datetimeoffset}(x) <- value

\method{tz}{datetimeoffset}(x)

tz(x) <- value
}
\arguments{
\item{x}{A \code{\link[=datetimeoffset]{datetimeoffset()}} object.}

\item{value}{Replacement value}
}
\description{
Methods for \code{\link[=datetimeoffset]{datetimeoffset()}} objects for the following
datetime component getter/setter functions from \code{lubridate}:
}
\details{
\itemize{
\item \code{\link[lubridate:date]{lubridate::date()}}
\item \code{\link[lubridate:year]{lubridate::year()}}
\item \code{\link[lubridate:month]{lubridate::month()}}
\item \code{\link[lubridate:day]{lubridate::mday()}} and its alias \code{\link[lubridate:day]{lubridate::day()}}
\item \code{\link[lubridate:hour]{lubridate::hour()}}
\item \code{\link[lubridate:minute]{lubridate::minute()}}
\item \code{\link[lubridate:second]{lubridate::second()}}
\item \code{\link[lubridate:tz]{lubridate::tz()}}
}
}
\examples{
library("lubridate", exclude = c("date", "force_tz", "tz<-", "with_tz"))
dt <- datetimeoffset(1984)
year(dt) <- 1918
year(dt)
month(dt) <- 11
month(dt)
day(dt) <- 11
day(dt)
hour(dt) <- 11
hour(dt)
minute(dt) <- 11
minute(dt)
second(dt) <- 11
second(dt)
if ("Europe/Paris" \%in\% OlsonNames()) {
  tz(dt) <- "Europe/Paris"
  tz(dt)
}

}
\seealso{
\link{getset_other}
}
