% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timezone.r
\name{mode_tz}
\alias{mode_tz}
\alias{mode_tz.datetime_offset}
\title{Get most common time zone}
\usage{
mode_tz(time, tzone = "", ...)

\method{mode_tz}{datetime_offset}(time, tzone = "", ...)
}
\arguments{
\item{time}{A datetime object.}

\item{tzone}{A timezone string to use for missing time zones.}

\item{...}{Ignored}
}
\value{
Timezone string
}
\description{
'mode_tz()' gets the most common time zone
in the datetime object.  If a tie we use the time zone used first.
Intended for use when coercing from a datetime object that supports
multiple heterogeneous time zones to a datetime object that
only supports one time zone
}
\examples{
  dt <- as_datetime_offset(Sys.time())
  print(mode_tz(dt))
  if (all(c("US/Pacific", "US/Eastern") \%in\% OlsonNames())) {
    dt <- as_datetime_offset("2020-01-01",
                             tz = c("US/Pacific", "US/Eastern"))
    print(mode_tz(dt))

    dt <- as_datetime_offset("2020-01-01",
                             tz = c("US/Pacific", "US/Eastern", NA_character_, NA_character_))
    print(mode_tz(dt))
  }
}
